:video_game: RigEditor: Remade wheels list + node info box
Removed "debug box" in top left corner - with DearIMGUI, adding ad-hoc debugging panels is so easy that pre-made debug box is not necessary anymore.

The "current mouse hovered node info" functionality was preserved - a new box was added to the bottom middle of screen. Only visible when mouse hovers a node.